all search data saved to database
