+++ .gitignore
@@ -2,3 +2,4 @@
 __pycache__/
 apis/__pycache__/
 apitestvenv/
+venv/
