+++ app.py
@@ -2,9 +2,13 @@ from flask import Flask, render_template, request  # NOT the same as requests
 import apis.spotify_api as spotify
 import apis.youtube_api as video
 import apis.ticketmaster_api as events
-import db.search_results_db as db
+import database.search_results_db as db
 
+DEBUG = True
 app = Flask(__name__)
+
+def store_info(artist_info):
+    db.database_connection(artist_info)
 
 @app.route('/') # home page
 def homepage():
@@ -19,15 +23,10 @@ def get_artist_info():
     artist_name = request.args.get('artist_name')
 
     artist_info = spotify.main(artist_name)
-    db.store_artist_info(artist_info)
+    store_info(artist_info)
     events_info = events.main(artist_name)
     music_video = video.main(f'{artist_info.artist} {artist_info.tracks[0]['title']}')
-    # send_to_db =
-    # db.store_data([
-    #     artist_info, events_info, music_video
-    # ])
 
-    # TODO add exception handling
     return render_template('search_result.html',
                            artist_name=artist_info.artist,
                            artist_img=artist_info.image_url,
